Revert "Country grouping"
