+++ src/pages/patientView/clinicalTrialMatch/utils/SelectValues.ts
@@ -14,51 +14,6 @@ export var recruitingValueNames: RecruitingStatus[] = [
     RecruitingStatus.Withdrawn,
 ];
 
-export var countriesGroups: { [key: string]: string[] } = {
-    'European Union member states': [
-        'Austria',
-        'Belgium',
-        'Bulgaria',
-        'Croatia',
-        'Cyprus',
-        'Czechia',
-        'Denmark',
-        'Estonia',
-        'Finland',
-        'France',
-        'Germany',
-        'Greece',
-        'Hungary',
-        'Ireland',
-        'Italy',
-        'Latvia',
-        'Lithuania',
-        'Luxembourg',
-        'Malta',
-        'Netherlands',
-        'Poland',
-        'Portugal',
-        'Romania',
-        'Slovakia',
-        'Slovenia',
-        'Spain',
-        'Sweden',
-    ],
-    'German-speaking': ['Germany', 'Austria', 'Switzerland'],
-    'Germany + neighbouring countries': [
-        'Germany',
-        'Austria',
-        'Belgium',
-        'Czech Republic',
-        'Denmark',
-        'France',
-        'Luxembourg',
-        'Netherlands',
-        'Poland',
-        'Switzerland',
-    ],
-};
-
 export var countriesNames: string[] = [
     'Afghanistan',
     'Albania',
